# server/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentOut": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          },
          "job_submitted_by": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "provision_type": {
            "type": "string"
          },
          "queues": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "restricted_to": {
            "additionalProperties": {},
            "type": "object"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "QueuesIn": {
        "additionalProperties": true,
        "properties": {},
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "detail": {
            "properties": {
              "<location>": {
                "properties": {
                  "<field_name>": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Testflinger API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/agents/queues": {
      "get": {
        "description": "Returns a dict of queue names and descriptions, ex:\n{\n\"some_queue\": \"A queue for testing\",\n\"other_queue\": \"A queue for something else\"\n}",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "example": {
                    "device001": "Queue for device001",
                    "some-queue": "some other queue"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Mapping of queue names and descriptions"
          }
        },
        "summary": "Get all advertised queues from this server.",
        "tags": [
          "V1"
        ],
        "x-permission-roles": [
          "admin",
          "manager",
          "contributor"
        ]
      },
      "post": {
        "description": "Some agents may want to advertise some of the queues they listen on so that\nthe user can check which queues are valid to use.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueuesIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Tell testflinger the queue names that are being serviced.",
        "tags": [
          "V1"
        ],
        "x-permission-roles": [
          "admin",
          "manager",
          "agent"
        ]
      }
    },
...
    "/v1/queues/{queue_name}/agents": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "queue_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Get the list of all data for agents listening to a specified queue.",
        "tags": [
          "V1"
        ],
        "x-permission-roles": [
          "admin",
          "manager",
          "contributor"
        ]
      }
    },
    "/v1/queues/{queue_name}/jobs": {
      "get": {
        "description": ":param queue_name\nString with the queue name where to perform the query.\n:return:\nJSON data with the jobs allocated to the specified queue.",
        "parameters": [
          {
            "in": "path",
            "name": "queue_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Get the jobs in a specified queue along with its state.",
        "tags": [
          "V1"
        ],
        "x-permission-roles": [
          "admin",
          "manager",
          "contributor"
        ]
      }
    },
...
  },
...
}