# server/src/testflinger/views.py
# Copyright (C) 2023-2024 Canonical
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.
#
"""Additional views not associated with the API."""

from datetime import datetime, timedelta, timezone
from http import HTTPStatus

import yaml
from apiflask import APIBlueprint
from flask import (
    current_app,
    make_response,
    render_template,
    request,
    session,
)
from markupsafe import Markup
from pygments import highlight as pygments_highlight
from pygments.formatters import HtmlFormatter
from pygments.lexers import BashLexer, YamlLexer

from testflinger import database
from testflinger.api.schemas import Job
from testflinger.database import mongo
from testflinger.logs import MongoLogHandler

views = APIBlueprint("testflinger", __name__, enable_openapi=False)


@views.before_request
def require_login():
    """Block UI views when OIDC is enabled and no user is logged in."""
    if current_app.oauth is None or request.endpoint == "testflinger.home":
        return None
    if "user" not in session:
        return make_response(
            render_template("401_unauthorized.html"),
            HTTPStatus.UNAUTHORIZED,
        )
    return None


# Fields the server assigns; never part of a resubmittable job definition.
_SERVER_MANAGED_JOB_FIELDS = frozenset({"job_id", "parent_job_id"})

# Submittable job-definition fields, derived from the Job schema in declaration
# order. Deriving (rather than hard-coding) means new submittable fields added
# to the schema appear in the copied YAML automatically; only server-managed
# fields need excluding above.
JOB_DEFINITION_FIELDS = tuple(
    name for name in Job().fields if name not in _SERVER_MANAGED_JOB_FIELDS
)


class _JobYamlDumper(yaml.SafeDumper):
    """YAML dumper that renders multiline strings as literal blocks."""


def _str_representer(dumper, data):
    """Represent multiline strings (e.g. test_cmds) as literal blocks."""
    style = "|" if "\n" in data else None
    return dumper.represent_scalar("tag:yaml.org,2002:str", data, style=style)


_JobYamlDumper.add_representer(str, _str_representer)


def _dump_yaml(data):
    """Dump a value to YAML using the literal-block-aware dumper."""
    return yaml.dump(
        data,
        Dumper=_JobYamlDumper,
        default_flow_style=False,
        sort_keys=False,
    )


def build_job_yaml(job_data):
    """Build a submittable job-definition YAML from stored job data."""
    definition = {
        field: job_data[field]
        for field in JOB_DEFINITION_FIELDS
        if field in job_data
    }
    return _dump_yaml(definition)


@views.app_template_filter("as_yaml")
def as_yaml(value):
    """Render a job-data section as YAML for display in the dashboard.

    Strings (e.g. the legacy ``provision_data: skip``) are shown as-is;
    mappings and lists are dumped as YAML instead of a Python repr.
    """
    if value is None or isinstance(value, str):
        return value
    return _dump_yaml(value).rstrip("\n")


# Built once and reused; nowrap emits only token <span>s (the <pre><code>
# wrapper lives in the template).
_HTML_FORMATTER = HtmlFormatter(nowrap=True)
_LEXERS = {"yaml": YamlLexer(), "bash": BashLexer()}


@views.app_template_filter("highlight")
def highlight(code, language):
    """Syntax-highlight code as safe HTML using Pygments.

    Returns Pygments token markup (content already HTML-escaped) for the
    given language; unknown languages fall back to the autoescaped string.
    """
    if not code:
        return ""
    lexer = _LEXERS.get(language)
    if lexer is None:
        return code
    rendered = pygments_highlight(code, lexer, _HTML_FORMATTER)
    # Pygments HTML-escapes the code content (only token markup is added), so
    # wrapping the result as safe markup does not introduce an XSS risk.
    return Markup(rendered.rstrip("\n"))  # noqa: S704


@views.route("/")
def home():
    """Home view."""
    return render_template("homepage.html")


@views.route("/agents")
def agents():
    """Agents view."""
    agent_info = database.get_agents()
    return render_template("agents.html", agents=agent_info)


@views.route("/agents/<agent_id>")
def agent_detail(agent_id):
    """Agent detail view."""
    default_start_date = (
        datetime.now(tz=timezone.utc) - timedelta(days=2)
    ).strftime("%Y-%m-%d")
    default_stop_date = datetime.now(tz=timezone.utc).strftime("%Y-%m-%d")

    start_date = request.args.get("start", default_start_date)
    stop_date = request.args.get("stop", default_stop_date)

    # Convert start and stop dates to datetime objects for the query
    start_datetime = datetime.strptime(start_date, "%Y-%m-%d").replace(
        tzinfo=timezone.utc
    )
    stop_datetime = datetime.strptime(stop_date, "%Y-%m-%d").replace(
        tzinfo=timezone.utc
    ) + timedelta(days=1)

    agent_info = mongo.db.agents.find_one({"name": agent_id})
    if not agent_info:
        response = make_response(
            render_template("agent_not_found.html", agent_id=agent_id)
        )
        response.status_code = HTTPStatus.NOT_FOUND
        return response

    restricted_queues = database.get_restricted_queues()
    client_permissions = database.get_restricted_queues_owners()

    agent_info["restricted_to"] = {
        queue: client_permissions.get(queue, [])
        if queue in restricted_queues
        else []
        for queue in agent_info.get("queues", [])
    }

    queue_info = []
    for queue_name in agent_info.pop("queues", []):
        queue_data = mongo.db.queues.find_one({"name": queue_name})
        if not queue_data:
            # If it's not an advertised queue, create some dummy data
            queue_data = {"description": ""}
        queue_data["name"] = queue_name
        queue_data["numjobs"] = database.get_num_incomplete_jobs_on_queue(
            queue_name
        )
        queue_info.append(queue_data)

    agent_info["queues"] = queue_info

    # We want to include the start/stop dates so that default values
    # can be filled in for the date pickers
    agent_info["start"] = start_date
    agent_info["stop"] = stop_date

    agent_info["provision_log"] = database.get_provision_log(
        agent_id,
        start_datetime=start_datetime,
        stop_datetime=stop_datetime,
    )

    # Enrich provision log entries with the submitted_by from the
    # corresponding job.
    job_ids = {
        entry["job_id"]
        for entry in agent_info["provision_log"]
        if entry.get("job_id")
    }
    if job_ids:
        submitted_by_map = {
            doc["job_id"]: doc.get("submitted_by")
            for doc in mongo.db.jobs.find(
                {"job_id": {"$in": list(job_ids)}},
                {"job_id": 1, "submitted_by": 1, "_id": 0},
            )
        }
        for entry in agent_info["provision_log"]:
            entry.setdefault(
                "submitted_by", submitted_by_map.get(entry.get("job_id"))
            )

    if agent_info["provision_log"]:
        agent_info["provision_success_rate"] = int(
            100
            * len(
                [
                    entry
                    for entry in agent_info["provision_log"]
                    if entry["exit_code"] == 0
                ]
            )
            / len(agent_info["provision_log"])
        )
    else:
        # Avoid division by zero
        agent_info["provision_success_rate"] = 0

    return render_template("agent_detail.html", agent=agent_info)


@views.route("/jobs")
def jobs():
    """Jobs view."""
    jobs_data = mongo.db.jobs.find({}, sort=[("created_at", -1)])
    return render_template("jobs.html", jobs=jobs_data)


@views.route("/jobs/<job_id>")
def job_detail(job_id):
    """Job detail view."""
    job_data = database.get_job(job_id)
    if not job_data:
        response = make_response(
            render_template("job_not_found.html", job_id=job_id),
            HTTPStatus.NOT_FOUND,
        )
        return response

    if result_data := job_data.get("result_data"):
        if not any(
            key.endswith(("_output", "_serial")) for key in result_data.keys()
        ):
            log_handler = MongoLogHandler(mongo)
            log_handler.format_logs_as_results(job_id, result_data)

    job_data["agent_name"] = job_data.get("result_data", {}).get("agent_id")
    job_yaml = build_job_yaml(job_data.get("job_data", {}))
    return render_template("job_detail.html", job=job_data, job_yaml=job_yaml)


@views.route("/queues")
def queues():
    """
    Queues view.

    Render a view with all known queues, descriptions if known, and the number
    of jobs in each.
    """
    queue_data = queues_data()
    return render_template("queues.html", queues=queue_data)


def queues_data():
    """Generate data for the queues view, this makes testing easier."""
    # First, get all the advertised queues with descriptions
    queue_data = list(
        mongo.db.queues.find(
            projection={"_id": 0, "name": 1, "description": 1}
        )
    )

    # Get all the queues the agents say they are listening to from agent data
    agent_data = mongo.db.agents.find({}, {"_id": 0, "queues": 1})
    agent_queues_set = {
        queue for agent in agent_data for queue in agent.get("queues", [])
    }
    #    queue for agent in agent_data for queue in agent["queues"]
    advertised_queues_set = {queue["name"] for queue in queue_data}

    # Only keep the ones that weren't also in the advertised queues
    unique_queues_from_agents = agent_queues_set - advertised_queues_set
    for queue_name in unique_queues_from_agents:
        queue_data.append(
            {"name": queue_name, "description": "", "numjobs": 0}
        )

    # Get job counts for each queue
    for queue in queue_data:
        queue["numjobs"] = database.get_num_incomplete_jobs_on_queue(
            queue["name"]
        )
    return queue_data


@views.route("/queues/<queue_name>")
def queue_detail(queue_name):
    """Queue detailed view."""
    queue_data = mongo.db.queues.find_one({"name": queue_name})
    if not queue_data:
        # If it's not an advertised queue, create some dummy data
        queue_data = {"name": queue_name, "description": "No description"}

    # Find all the jobs active jobs in this queue
    job_data = mongo.db.jobs.find(
        {
            "job_data.job_queue": queue_name,
            "result_data.job_state": {
                "$nin": ["complete", "completed", "cancelled"]
            },
        }
    )

    # Get the percentiles of wait times for this queue
    wait_times = database.get_queue_wait_times([queue_name])
    try:
        wait_times = wait_times[0]["wait_times"]
    except (IndexError, KeyError):
        wait_times = []
    queue_percentile_data = database.calculate_percentiles(wait_times)

    # Convert the wait times to human-readable strings
    for key, value in queue_percentile_data.items():
        queue_percentile_data[key] = seconds_to_hms(value)

    agents_data = database.get_agents(queue=queue_name)

    return render_template(
        "queue_detail.html",
        queue=queue_data,
        jobs=job_data,
        queue_percentile_data=queue_percentile_data,
        agents=agents_data,
    )


def seconds_to_hms(seconds: float) -> str:
    """Convert seconds to a human-readable string."""
    seconds = int(seconds)
    hours = seconds // 3600
    minutes = (seconds % 3600) // 60
    seconds = seconds % 60
    return f"{hours:02d}h {minutes:02d}m {seconds:02d}s"